+++ contracts/solana-types/src/types.rs
@@ -15,7 +15,7 @@ pub struct TokenInfo {
 #[derive(Debug, BorshSerialize, BorshDeserialize)]
 pub struct CreateGameAccountParams {
     pub title: String,
-    pub max_players: u8,
+    pub max_players: u16,
     pub min_deposit: u64,
     pub max_deposit: u64,
     pub data: Vec<u8>,
